+++ packages/nip47/lib/nip47.dart
@@ -13,4 +13,10 @@ export 'src/data/models/response.dart';
 export 'src/data/models/tlv_record.dart';
 export 'src/data/models/transaction.dart';
 
+export 'src/enums/bitcoin_network.dart';
+export 'src/enums/error_code.dart';
+export 'src/enums/event_kind.dart';
+export 'src/enums/method.dart';
+export 'src/enums/transaction_type.dart';
+
 export 'src/constants.dart';