+++ packages/nip47/test/nip47_test.dart
@@ -1,6 +1,5 @@
 import 'package:nip01/nip01.dart';
 import 'package:nip47/nip47.dart';
-import 'package:nip47/src/enums/method.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -8,13 +7,17 @@ void main() {
     'adds a connection',
     () async {
       const relayUrl = 'wss://nostr2.daedaluslabs.io';
+      final nip01Repository = Nip01RepositoryImpl(
+        relayClientsManager: RelayClientsManagerImpl([relayUrl]),
+      );
       final nostrKeyPair = KeyPair.generate();
-      final nwcWallet = Wallet(
-        relayUrl: relayUrl,
+      final nwcWallet = WalletServiceImpl(
         walletKeyPair: nostrKeyPair,
+        nip01repository: nip01Repository,
       );
 
       final connection = await nwcWallet.addConnection(
+        relayUrl: relayUrl,
         permittedMethods: [
           Method.getInfo,
           Method.getBalance,