+++ CHANGELOG.md
@@ -3,6 +3,101 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2024-10-05
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`nip24` - `v0.0.3`](#nip24---v003)
+
+---
+
+#### `nip24` - `v0.0.3`
+
+
+## 2024-10-05
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`nip19` - `v0.0.3`](#nip19---v003)
+
+---
+
+#### `nip19` - `v0.0.3`
+
+
+## 2024-10-05
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`nip06` - `v0.0.3`](#nip06---v003)
+
+---
+
+#### `nip06` - `v0.0.3`
+
+
+## 2024-10-05
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`nip47` - `v0.1.0`](#nip47---v010)
+
+---
+
+#### `nip47` - `v0.1.0`
+
+
+## 2024-10-05
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`nip01` - `v0.1.0`](#nip01---v010)
+
+---
+
+#### `nip01` - `v0.1.0`
+
+
 ## 2024-09-30
 
 ### Changes