+++ packages/nip01/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0
+
+- Abstract away relay connection and reconnect logic.
+- Add `Nip01Repository` class.
+- Add `RelayClientsManager` class.
+- Change `RelayCommunication` to `RelayClient` class.
+- Change `RelayConnectionProvider` to `RelayStreamProvider` class.
+
 ## 0.0.2
 
 - Update repository URL in pubspec.yaml.