+++ packages/nip06/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+- Update `nip01` dependency to `0.1.0`.
+
 ## 0.0.2
 
 - Update repository URL in pubspec.yaml.